LineNumberExtensionHelper to LineNumberExtensionContext
