+++ MarkdigEngine/MarkdigEngine.csproj
@@ -84,6 +84,7 @@
     <Compile Include="Extensions\Inclusion\InclusionInline\InclusionInline.cs" />
     <Compile Include="Extensions\Inclusion\InclusionInline\InclusionInlineParser.cs" />
     <Compile Include="Extensions\LineNumber\LineNumberExtension.cs" />
+    <Compile Include="Extensions\LineNumber\LineNumberExtensionContext.cs" />
     <Compile Include="MarkdigMarked.cs" />
     <Compile Include="Extensions\CodeSnippet\CodeRange.cs" />
     <Compile Include="Extensions\CodeSnippet\HtmlCodeSnippetRenderer.cs" />
